FLOW-797 code duplication reduced
