+++ stories/link-test.stories.ts
@@ -1,10 +1,7 @@
 import { Story } from "@storybook/web-components";
 import { html } from "lit-html";
-import {
-  LineageNodeElement,
-  LineageNodeLinks,
-  LineageNodes,
-} from "@cldcvr/flow-lineage/src";
+import { LineageNodeLinks, LineageNodes } from "@cldcvr/flow-lineage/src";
+import { nodeTemplate, childrenNodeTemplate } from "./nodeTemplates";
 
 // export default {
 //   title: "Debug/Links",
@@ -169,44 +166,8 @@ const Template: Story<unknown> = (args: any) => {
 export const basic = Template.bind({});
 
 basic.args = {
-  ["node-template"]: function (node: LineageNodeElement) {
-    return html`<f-div
-      state=${node.fData?.state}
-      width="100%"
-      height="100%"
-      padding="small"
-      align="top-left"
-      variant="curved"
-      gap="small"
-      ${node.fChildren && !node.fHideChildren
-        ? 'border="small solid default bottom"'
-        : ""}
-    >
-      <f-pictogram
-        variant="circle"
-        source="${node.fData?.fullName}"
-      ></f-pictogram>
-      <f-div direction="column">
-        <f-text size="small" ellipsis>${node.fData?.fullName}</f-text>
-        <f-text size="small" ellipsis> x : ${node.x} , y : ${node.y}</f-text>
-      </f-div>
-      ${node.childrenToggle}
-    </f-div>`;
-  },
-  ["children-node-template"]: function (node: LineageNodeElement) {
-    return html`<f-div
-      state="secondary"
-      width="100%"
-      height="100%"
-      padding="none medium"
-      align="middle-left"
-      gap="small"
-      border="small solid default bottom"
-    >
-      <f-icon source="${node.fData?.icon}" size="small"></f-icon>
-      <f-text size="small" ellipsis>${node.fData?.title}</f-text>
-    </f-div>`;
-  },
+  ["node-template"]: nodeTemplate,
+  ["children-node-template"]: childrenNodeTemplate,
 };
 
 // let updateCounter = 0;
